test: clean navigation imports
